fix trajectoire smooth

+++ ObjetCeleste.java
@@ -10,17 +10,18 @@ public abstract class ObjetCeleste {
     //
     public double masse, vx, vy;
     public int r;
-    protected int x, y;
+    protected double x, y;
     protected Image disp;
     protected Image imageObjet;
     protected HUD hud;
     protected double zoomCoeff;
     protected int rZoom;
-    protected int xZ,yZ;
+    protected double xZ,yZ;
     public boolean atome[] = new boolean [4];
 
 
-    public ObjetCeleste(double m, double vitx, double vity, int ax, int ay, Image i, int rayon, HUD hud, boolean [] a) {
+    public ObjetCeleste(double m, double vitx, double vity, double ax, double ay, Image i, int rayon, HUD hud, boolean [] a) {
+        
         masse = m;
         x = ax;
         y = ay;
@@ -36,7 +37,6 @@ public abstract class ObjetCeleste {
         for (int j =0 ; j<a.length ; j++ ) {
             atome[j] = a[j];
         }
-
 
         //bug : la linked list n'est pas dans le constructeur on ne peut pas y avoir acces depuis ici il faut faire en sort de créer la methode update dans space.java
         //pour avoir acces a tous les objets, voir aussi au niveau du foreach plus simple et propre pour parcourir une linked list
@@ -73,11 +73,11 @@ public abstract class ObjetCeleste {
 
     //accesseurs
 
-    public int GetX() {
+    public double GetX() {
         return this.x;
     }
 
-    public int GetY() {
+    public double GetY() {
         return this.y;
     }
 
@@ -124,7 +124,7 @@ public abstract class ObjetCeleste {
     public void InitializeTemp(double t){
         hud.InitializeTemp(t);
     }
-    
+
     public void SetTemp(double t) {
         hud.SetTemp(t);
     }
