+++ Soleil.java
@@ -1,8 +1,7 @@
-import java.awt.Color;
 import java.awt.Graphics;
 import java.awt.Image;
-
 import java.io.File;
+
 import javax.imageio.ImageIO;
 
 /**
@@ -10,7 +9,7 @@ import javax.imageio.ImageIO;
  */
 public class Soleil extends ObjetCeleste {
 
-    public Soleil(double m, int ax, int ay, Image i, int rayon, HUD hud) {
+    public Soleil(double m, double ax, double ay, Image i, int rayon, HUD hud) {
         super(m, 0, 0, ax, ay, i, rayon, hud, new boolean[4]); //un soleil est une étoile fixe avec vitesse = 0
         
         hud.InitializeTemp(1000);
@@ -25,7 +24,7 @@ public class Soleil extends ObjetCeleste {
     }
 
     public void paint(Graphics g){
-        g.drawImage(disp, this.xZ - disp.getWidth(null) / 2, this.yZ - disp.getHeight(null) / 2, null);
+        g.drawImage(disp, (int)this.xZ - disp.getWidth(null) / 2, (int)this.yZ - disp.getHeight(null) / 2, null);
     }
 
     @Override
